+++ app/(club)/club/management/activity/accounting/page.tsx
@@ -2,10 +2,10 @@ import MainSection from "./content/mainSection";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
-  title: "회계 내역 | 수입과 지출을 관리하세요",
+  title: "회계 내역 | 아리아리",
   description: "회계 내역을 확인하고 투명하게 관리할 수 있습니다.",
   openGraph: {
-    title: "회계 내역 | 수입과 지출을 관리하세요",
+    title: "회계 내역 | 아리아리",
     description: "회계 내역을 확인하고 투명하게 관리할 수 있습니다.",
     url: "https://ariari.com/club/management/activity/accounting",
     siteName: "아리아리",
