+++ app/(club)/club/management/activity/notice/page.tsx
@@ -2,10 +2,10 @@ import MainSection from "./content/mainSection";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
-  title: "공지사항 | 주요 소식을 전달합니다",
+  title: "공지사항 | 아리아리",
   description: "공지사항을 등록하고 멤버들에게 전달할 수 있는 공간입니다.",
   openGraph: {
-    title: "공지사항 | 주요 소식을 전달합니다",
+    title: "공지사항 | 아리아리",
     description: "공지사항을 등록하고 멤버들에게 전달할 수 있는 공간입니다.",
     url: "https://ariari.com/club/management/activity",
     siteName: "아리아리",
