+++ app/(recruitment)/recruitment/detail/content/recruitmentDetail.tsx
@@ -1,20 +1,35 @@
-"use client";
-
 import React, { useEffect, useState } from "react";
 import { useSearchParams } from "next/navigation";
-
 import ClubInfo from "./clubInfo";
 import ClubActivities from "./clubActivities";
-
-import { RecruitmentData, RecruitmentResponse } from "@/types/recruitment";
-
-import {
-  getClubRecruitmentList,
-  getRecruitmentDetail,
-} from "@/api/recruitment/api";
+import { RecruitmentData } from "@/types/recruitment";
+import { getClubRecruitmentList } from "@/api/recruitment/api";
 import Loading from "@/components/feedback/loading";
 import ErrorNotice from "@/components/feedback/error";
 import { useRecruitmentDetailQuery } from "@/hooks/recruitment/useRecruitmentDetailQuery";
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "모집공고 상세 | 아리아리",
+  description:
+    "새로운 멤버를 모집합니다! 활동 내용, 지원 자격, 모집 일정 등 상세 정보를 확인하세요.",
+  openGraph: {
+    title: "모집공고 상세 | 아리아리",
+    description:
+      "새로운 멤버를 모집합니다! 활동 내용, 지원 자격, 모집 일정 등 상세 정보를 확인하세요.",
+    url: "https://www.ariari.kr/recruitment",
+    siteName: "아리아리",
+    images: [
+      {
+        url: "/logo.svg",
+        width: 1200,
+        height: 630,
+        alt: "아리아리",
+      },
+    ],
+    type: "website",
+  },
+};
 
 const RecruitmentDetail = () => {
   const params = useSearchParams();
