+++ app/(user)/user/interestRecruitment/page.tsx
@@ -2,11 +2,11 @@ import MainSection from "./content/mainSection";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
-  title: "관심 모집공고 | 찜한 모집 정보를 모아보세요",
+  title: "관심 모집공고 | 아리아리",
   description:
     "찜한 동아리의 모집 공고를 한눈에 모아보고 쉽게 관리할 수 있습니다.",
   openGraph: {
-    title: "관심 모집공고 | 찜한 모집 정보를 모아보세요",
+    title: "관심 모집공고 | 아리아리",
     description:
       "찜한 동아리의 모집 공고를 한눈에 모아보고 쉽게 관리할 수 있습니다.",
     url: "https://ariari.com/user/interestRecruitment",
